tools/baremetal-ci/requirements.txt
fastapi
uvicorn
python-multipart
pymongo
requests
coloredlogs
python-dotenv
